refactor(platform): single-source the node:fs filesystem fallbacks
isSymlink, realPath, and readFileChunk were byte-identical in both
FileSystemProvider implementations: the VS Code adapter deliberately
bypasses vscode.workspace.fs for these (no chunked read, and the
symbolic-link bit is unreliable across platforms) and falls back to
node:fs, exactly as the Node provider does. The "why" comment for
isSymlink lived in only one of the two copies.

Move the three bodies into one src/platform/defaults/nodeFsOps.ts that
both providers delegate to, so the Node-fs fallback surface and its
rationale have a single owner. Behavior-preserving: identical bodies,
read-only fs ops; readFile stays per-provider (it genuinely differs).
