+++ src/platform/defaults/nodeFilesystem.ts
@@ -13,6 +13,7 @@ import {
   type FileSystemProvider,
   type FileStat,
 } from '../interfaces/filesystem';
+import * as nodeFsOps from './nodeFsOps';
 
 /**
  * Resolve the target type of a symlink, producing combined bitmasks
@@ -92,33 +93,15 @@ export const nodeFilesystem: FileSystemProvider = {
     };
   },
 
-  async isSymlink(target: string): Promise<boolean> {
-    const lstats = await fs.promises.lstat(target);
-    return lstats.isSymbolicLink();
-  },
+  isSymlink: nodeFsOps.isSymlink,
 
-  async realPath(target: string): Promise<string> {
-    return fs.promises.realpath(target);
-  },
+  realPath: nodeFsOps.realPath,
 
   async readFile(target: string): Promise<Uint8Array> {
     return fs.promises.readFile(target);
   },
 
-  async readFileChunk(
-    target: string,
-    offset: number,
-    length: number,
-  ): Promise<Uint8Array> {
-    const handle = await fs.promises.open(target, 'r');
-    try {
-      const buffer = Buffer.alloc(length);
-      const { bytesRead } = await handle.read(buffer, 0, length, offset);
-      return buffer.subarray(0, bytesRead);
-    } finally {
-      await handle.close();
-    }
-  },
+  readFileChunk: nodeFsOps.readFileChunk,
 
   async writeFile(target: string, content: Uint8Array): Promise<void> {
     await fs.promises.writeFile(target, content);
